+++ init.lua
@@ -63,7 +63,7 @@ local function enable(self,config,stage)
       ui.resetButtons()
     end)
     observe(native.addr(0x46b358),6,function(registers)
-      if registers.EBP==61 and not engine.loading then recorder:reset(); ui.resetButtons() end
+      recorder:onMenuView(registers.EBP)
     end)
     observe(native.addr(0x495337),6,function()
       if not engine.loading then recorder:reset() end